+++ service/history/transferQueueFactory.go
@@ -98,7 +98,7 @@ func NewTransferQueueFactory(
 					params.Config.PersistenceMaxQPS,
 					transferQueuePersistenceMaxRPSRatio,
 				),
-				params.Config.QueueMaxReaderCount(),
+				int64(params.Config.QueueMaxReaderCount()),
 			),
 		},
 	}
